+++ src/frontend/apps/standalone_site/src/features/ClassRooms/index.tsx
@@ -1,0 +1,1 @@
+export { default as ClassRooms } from './components/ClassRooms';