+++ src/frontend/apps/standalone_site/src/routes/AppRoutes.spec.tsx
@@ -29,5 +29,11 @@ describe('<AppRoutes />', () => {
     await waitFor(() => {
       expect(screen.getByText(/My favorite/i)).toBeInTheDocument();
     });
+
+    fireEvent.click(screen.getByRole(/menuitem/i, { name: /Classrooms/i }));
+
+    await waitFor(() => {
+      expect(screen.getByText(/Classrooms/i)).toBeInTheDocument();
+    });
   });
 });