+++ src/frontend/apps/standalone_site/src/utils/date.ts
@@ -1,0 +1,6 @@
+import { DateTime } from 'luxon';
+
+export const localDate = (dateNotFormated: string, locale: string) => {
+  const dt = DateTime.fromISO(dateNotFormated);
+  return dt.isValid ? dt.setLocale(locale).toFormat('D  ·  tt') : null;
+};